pyproject.toml
[project]
name = "gitbridge"
version = "0.1.0"
description = "A Git fetcher container that clones two repos with flexible auth"
authors = [{ name = "Paul Dresch" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython>=3.1.45",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "structlog>=25.4.0",
    "tenacity>=9.1.2",
    "typer>=0.16.1",
]
